Completing tray icon implementation on Windows
