+++ lib/tray/tray.h
@@ -220,54 +220,13 @@ static void tray_exit() { objc_msgSend(app, sel_registerName("terminate:"), app)
 
 #include <shellapi.h>
 
-#define WM_TRAY_CALLBACK_MESSAGE (WM_USER + 1)
-#define WC_TRAY_CLASS_NAME "TRAY"
+#define WM_TRAY_CALLBACK_MESSAGE (WM_USER + 2)
+#define WM_TRAY_PASS_MENU_REF (WM_USER + 1)
 #define ID_TRAY_FIRST 1000
 
-static WNDCLASSEX wc;
 static NOTIFYICONDATA nid;
 static HWND hwnd;
 static HMENU hmenu = NULL;
-
-static LRESULT CALLBACK _tray_wnd_proc(HWND hwnd, UINT msg, WPARAM wparam,
-                                       LPARAM lparam) {
-  switch (msg) {
-  case WM_CLOSE:
-    DestroyWindow(hwnd);
-    return 0;
-  case WM_DESTROY:
-    PostQuitMessage(0);
-    return 0;
-  case WM_TRAY_CALLBACK_MESSAGE:
-    if (lparam == WM_LBUTTONUP || lparam == WM_RBUTTONUP) {
-      POINT p;
-      GetCursorPos(&p);
-      SetForegroundWindow(hwnd);
-      WORD cmd = TrackPopupMenu(hmenu, TPM_LEFTALIGN | TPM_RIGHTBUTTON |
-                                           TPM_RETURNCMD | TPM_NONOTIFY,
-                                p.x, p.y, 0, hwnd, NULL);
-      SendMessage(hwnd, WM_COMMAND, cmd, 0);
-      return 0;
-    }
-    break;
-  case WM_COMMAND:
-    if (wparam >= ID_TRAY_FIRST) {
-      MENUITEMINFO item;
-      memset(&item, 0, sizeof(item));
-      item.cbSize = sizeof(MENUITEMINFO);
-      item.fMask = MIIM_ID | MIIM_DATA;
-      if (GetMenuItemInfo(hmenu, wparam, FALSE, &item)) {
-        struct tray_menu *menu = (struct tray_menu *)item.dwItemData;
-        if (menu != NULL && menu->cb != NULL) {
-          menu->cb(menu);
-        }
-      }
-      return 0;
-    }
-    break;
-  }
-  return DefWindowProc(hwnd, msg, wparam, lparam);
-}
 
 static HMENU _tray_menu(struct tray_menu *m, UINT *id) {
   HMENU hmenu = CreatePopupMenu();
@@ -302,16 +261,8 @@ static HMENU _tray_menu(struct tray_menu *m, UINT *id) {
 }
 
 static int tray_init(struct tray *tray) {
-  memset(&wc, 0, sizeof(wc));
-  wc.cbSize = sizeof(WNDCLASSEX);
-  wc.lpfnWndProc = _tray_wnd_proc;
-  wc.hInstance = GetModuleHandle(NULL);
-  wc.lpszClassName = WC_TRAY_CLASS_NAME;
-  if (!RegisterClassEx(&wc)) {
-    return -1;
-  }
 
-  hwnd = CreateWindowEx(0, WC_TRAY_CLASS_NAME, NULL, 0, 0, 0, 0, 0, 0, 0, 0, 0);
+  hwnd = FindWindow("Neutralinojs_webview", NULL);
   if (hwnd == NULL) {
     return -1;
   }
@@ -324,8 +275,11 @@ static int tray_init(struct tray *tray) {
   nid.uFlags = NIF_ICON | NIF_MESSAGE;
   nid.uCallbackMessage = WM_TRAY_CALLBACK_MESSAGE;
   Shell_NotifyIcon(NIM_ADD, &nid);
-
   tray_update(tray);
+  
+  // Send menu ref to webview's Window event listener,
+  // because WM_TRAY_CALLBACK_MESSAGE needs it
+  SendMessage(hwnd, WM_TRAY_PASS_MENU_REF, (WPARAM)hmenu, 0); 
   return 0;
 }
 
@@ -360,7 +314,7 @@ static void tray_update(struct tray *tray) {
   }
 }
 
-static void tray_exit() {
+static void tray_exit(HMENU hmenu) {
   Shell_NotifyIcon(NIM_DELETE, &nid);
   if (nid.hIcon != 0) {
     DestroyIcon(nid.hIcon);
@@ -369,7 +323,6 @@ static void tray_exit() {
     DestroyMenu(hmenu);
   }
   PostQuitMessage(0);
-  UnregisterClass(WC_TRAY_CLASS_NAME, GetModuleHandle(NULL));
 }
 #else
 #error Please define TRAY_WINAPI, TRAY_APPINDICATOR or TRAY_APPKIT before including this file.
